requirements.txt
python-binance
pandas
python-dotenv
numpy
pandas-ta
